Revised Cleaning + Modeling Code (#2)
* Types and Abilities Cleaning

Added one hot encoding for the types and abilities variables in the original dataset.

* Corrections to the compiled code

Made corrections to bugs that occurred when compiling the code, as well as added a line to drop the raw columns after we've extracted info from them

* Code Cleaning + New Abilities Names

A partial update of code cleaning + a new list of names for abilities that removes the spaces

* Separate Files for Abilities Lists

Created lists stored in text files for the abilities lists so they didn't have to take up so much visual space in the main R file

* Cleaned the Data Cleaning Code

Cleaned and commented the data cleaning code.  Also fixed a mistake where the moves power summation loop was arbitrarily only running for 100 observations.

* Manually Added Missing Moves

The moves dataset is missing a bunch of moves, presumably due to name errors from the data scraping process. As such we've manually added them to the dataset here so they can be referenced later in analysis.

* Combined New Cleaning Code and Modeling Code

Combined the revised version of the cleaning code with the random forest + GLM code. Made some minor cleaning tweaks to the forest/GLM code but nothing major on that half.